+++ src/main/java/com/mongle/api/service/CommentService.java
@@ -3,21 +3,17 @@ package com.mongle.api.service;
 import com.mongle.api.domain.Comment;
 import com.mongle.api.domain.Post;
 import com.mongle.api.domain.User;
-import com.mongle.api.domain.dto.CommentReqDto;
-import com.mongle.api.domain.dto.CommentResDto;
-import com.mongle.api.domain.enums.Status;
+import com.mongle.api.dto.comment.CommentReqDto;
+import com.mongle.api.dto.comment.CommentResDto;
 import com.mongle.api.exception.GeneralException;
 import com.mongle.api.repository.CommentRepository;
-import com.mongle.api.repository.PostRepository;
 import com.mongle.api.response.code.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 @Service
 @Slf4j
